@@ -1,7 +1,8 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+	<TargetFramework>net6.0</TargetFramework>
+	<ImplicitUsings>enable</ImplicitUsings>
 
     <PackageId>Cofoundry.Plugins.Mail.SendGrid</PackageId>
     <Description>Cofoundry mail service implemention for SendGrid.</Description>
@@ -11,12 +12,12 @@
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
-    <DocumentationFile>bin\Release\netcoreapp3.1\Cofoundry.Plugins.Mail.SendGrid.xml</DocumentationFile>
+    <DocumentationFile>bin\Release\net6.0\Cofoundry.Plugins.Mail.SendGrid.xml</DocumentationFile>
     <NoWarn>1701;1702;1705;1591</NoWarn>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Cofoundry.Core" Version="0.10.0" />
+    <PackageReference Include="Cofoundry.Core" Version="0.11.1" />
     <PackageReference Include="Sendgrid" Version="9.26.0" />
   </ItemGroup>
 
